app-history/navigate/navigate-beforeunload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let iframe_exception = i.contentWindow.DOMException;
  let navigate_event_count = 0;
  i.contentWindow.appHistory.onnavigate = () => navigate_event_count++;
  i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func('onnavigatesuccess should not be called');
  i.contentWindow.appHistory.onnavigateerror = t.unreached_func('onnavigateerror should not be called');
  let onbeforeunload_promise;
  i.contentWindow.onbeforeunload = t.step_func(() => {
    onbeforeunload_promise = i.contentWindow.appHistory.navigate("#").catch(e => {
      assert_equals(e.constructor, i.contentWindow.DOMException);
    });
  });
  i.contentWindow.appHistory.navigate("?1");
  await onbeforeunload_promise;
  assert_equals(navigate_event_count, 1);
}, "Navigate inside onbeforeunload");
</script>
